Updates
Traying to implement ML but error in d_training. In this bit the results of the contours append and in another list I append the classes and print them as arrays. Still working as the contours are not accurate, i need to play with the values to make the camara recognize de shapes with the yellow cards.