Working 2.0
By now the prediction is working better using the camera stream as I changed some 'filters' making them match with the ones applied for the extraction of the contours from the pictures. The colour recognition works but it can vary depending the light and give wrong results. But when I apply it in the stream with the contour recognition gives me always blue. I need to play with them to see if there is some improvement.

PENDING:
- GUI
- Double check colour recognition
-Confirm at daylight that the contour recognitions works properly
-Readme file
-Submission